Move dependencies page to components directory
