+++ zipkin-lens/src/components/App/App.tsx
@@ -26,8 +26,8 @@ import { useTitle } from 'react-use';
 import { ThemeProvider } from 'styled-components';
 
 import Layout from './Layout';
+import DependenciesPage from '../DependenciesPage';
 import DiscoverPage from '../DiscoverPage';
-import DependenciesPage from '../../dependencies/DependenciesPage';
 import TracePage from '../TracePage';
 import { UiConfig, UiConfigConsumer } from '../UiConfig';
 import configureStore from '../../store/configure-store';
