+++ zipkin-lens/src/components/DependenciesPage/DependenciesGraph.tsx
@@ -30,7 +30,7 @@ import { AutoSizer } from 'react-virtualized';
 import Edge from './Edge';
 import NodeDetailData from './NodeDetailData';
 import VizceralWrapper from './VizceralWrapper';
-import Dependencies from '../models/Dependencies';
+import Dependencies from '../../models/Dependencies';
 
 // These filter functions use any type because they are passed directly to untyped JS code.
 const filterConnections = (object: any, value: any) => {
