+++ zipkin-lens/src/components/DependenciesPage/NodeDetailData.tsx
@@ -33,7 +33,7 @@ import { withRouter, RouteComponentProps } from 'react-router-dom';
 import { PieChart, Pie, Cell, ResponsiveContainer } from 'recharts';
 
 import Edge from './Edge';
-import { selectServiceColor } from '../colors';
+import { selectServiceColor } from '../../colors';
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
